// src/app/types/weapon.type.ts
export enum WeaponType {
	GreatSword = 'GreatSword',
	LongSword = 'LongSword',
	SwordAndShield = 'SwordAndShield',
	DualBlades = 'DualBlades',
	Hammer = 'Hammer',
	HuntingHorn = 'HuntingHorn',
	Lance = 'Lance',
	Gunlance = 'Gunlance',
	SwitchAxe = 'SwitchAxe',
	ChargeBlade = 'ChargeBlade',
	InsectGlaive = 'InsectGlaive',
	LightBowgun = 'LightBowgun',
	HeavyBowgun = 'HeavyBowgun',
	Bow = 'Bow'
}